TgBot/requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
APScheduler>=3.10.0 